+++ src/lib.rs
@@ -106,13 +106,14 @@ pub mod rng;
 /// Mock network
 #[cfg(feature = "mock_base")]
 pub use self::{
-    chain::SectionKeyShare,
     consensus::generate_bls_threshold_secret_key,
     messages::{AccumulatingMessage, Message, PlainMessage, Variant},
     network_params::NetworkParams,
     relocation::Overrides as RelocationOverrides,
     routing_table::delivery_group_size,
-    section::{elders_info_for_test, quorum_count, section_proof_slice_for_test, MIN_AGE},
+    section::{
+        elders_info_for_test, quorum_count, section_proof_slice_for_test, SectionKeyShare, MIN_AGE,
+    },
     xor_space::Xorable,
 };
