+++ src/section/section_keys.rs
@@ -1,0 +1,59 @@
+// Copyright 2020 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+use super::elders_info::EldersInfo;
+use crate::{consensus::DkgResult, id::PublicId};
+
+/// The secret share of the section key.
+#[derive(Clone)]
+pub struct SectionKeyShare {
+    /// Index used to combine signature share and get PublicKeyShare from PublicKeySet.
+    pub index: usize,
+    /// Secret Key share
+    pub key: bls::SecretKeyShare,
+}
+
+impl SectionKeyShare {
+    /// Create a new share with associated share index.
+    #[cfg(any(test, feature = "mock_base"))]
+    pub const fn new_with_position(index: usize, key: bls::SecretKeyShare) -> Self {
+        Self { index, key }
+    }
+
+    /// create a new share finding the position wihtin the elders.
+    pub fn new(
+        key: bls::SecretKeyShare,
+        our_id: &PublicId,
+        new_elders_info: &EldersInfo,
+    ) -> Option<Self> {
+        Some(Self {
+            index: new_elders_info.member_ids().position(|id| id == our_id)?,
+            key,
+        })
+    }
+}
+
+/// All the key material needed to sign or combine signature for our section key.
+#[derive(Clone)]
+pub struct SectionKeys {
+    /// Public key set to verify threshold signatures and combine shares.
+    pub public_key_set: bls::PublicKeySet,
+    /// Secret Key share and index. None if the node was not participating in the DKG.
+    pub secret_key_share: Option<SectionKeyShare>,
+}
+
+impl SectionKeys {
+    pub fn new(dkg_result: DkgResult, our_id: &PublicId, new_elders_info: &EldersInfo) -> Self {
+        Self {
+            public_key_set: dkg_result.public_key_set,
+            secret_key_share: dkg_result
+                .secret_key_share
+                .and_then(|key| SectionKeyShare::new(key, our_id, new_elders_info)),
+        }
+    }
+}
